Resolves util/expand-automation conflicts; condenses cc tests into fewer files
