+++ e2e/cypress/fixtures/state/store.json
@@ -1,1 +1,11 @@
-{}
+{
+  "credentials": "{\"clientId\": \"sa-platform-e0000000-c7ccdc8b5f18\", \"clientSecret\": \"6bf5cd85-a4de-456d-b04a-2ddb6554d00a\"}",
+  "clientcredentials": "{\"clientId\": \"4FB68156-CFB7CDC2B12\", \"clientSecret\": \"56886533-e1ea-43ea-ba53-13db48e43744\", \"tokenEndpoint\": \"http://keycloak.localtest.me:9080/auth/realms/master/protocol/openid-connect/token\"}",
+  "clientidsecret": "{\"clientId\": \"C7FB670A-5F1D7E94913\", \"clientSecret\": \"852b5f14-9b19-442c-800b-d846845e489d\", \"tokenEndpoint\": \"http://keycloak.localtest.me:9080/auth/realms/master/protocol/openid-connect/token\"}",
+  "jwtkeypaircredentials": "{\"clientId\": \"4A4BD841-EC9533C28C6\", \"tokenEndpoint\": \"http://keycloak.localtest.me:9080/auth/realms/master/protocol/openid-connect/token\"}",
+  "jwksurlkeys": "{\n  \"keys\": [\n    {\n      \"kty\": \"RSA\",\n      \"kid\": \"9EYL-Yl5dqiO0wbCyMKodOI4fC4U4IVaIfkj2kW7cw4\",\n      \"use\": \"sig\",\n      \"alg\": \"RS256\",\n      \"e\": \"AQAB\",\n      \"n\": \"tYl-DJHVbqClDiNXlsag3FDY92GOG1p9XHdv8Hsjoz9X5lsY4wHCM9d5PEK-Q3Q-cY-YH4w0YbP6D4GqZoRSJbtcsCCgUS5hk2SPNI5Jsv-_RiuzInSrTKLMPgZKDiVoe5W-G5yhoUOmhtABBvrwppnfff5o2JEfdeak1V7zA7WUWJjR6olPxNt3SbAbz456UYMvD7sAW2ap8hDEmYxL_LKOhKb_P4Usoho7BIq0DhT5EoJ7Nnorf3tICfuzMQ6WB53cQaJOV8zKiap6JTQsHKMA5G1gr-onbKfk7om1kYmTJ8UV8or40z70xmeCDNsFPMe3pizh96Q3yXGxeUkjZw\",\n      \"d\": \"a4O4sX0VaOWfm5X25AdTceg395tBbJXVNRRPqZ9kZeS4DGfbf2alO8rbtILa6FxKNzuSHu0MgOj0qoZjA0BO3qXuqSlJiYrsozylS70RjP_Gqqpq93cdlM--IuXaqjIhhUAUTYogd7_3OlSqpxWrvR_Yq2gndEvO1Cb7bJtFVeYOHIIrVw-spLlz5a-71V_ePNAJveGsf0z5Zxrn9Eq5URFDIllR9Rhn9XdhY8zScwYCXSx2lEoEXWu44oFXkXfx-iICtYSgfIk0GxJdqU6xgyzXsa-O1PnF9zMuG5nVia_wA47a1boWLROdhgdPN7WSaFBVtYiEAXtrxpnp7KXZQQ\",\n      \"p\": \"7ZOPCHMsQeoaPRFEua-5B41_uiNBrJjPzmDaHFDQkU4KeJtpt94wYgGr_s9ImhyzopmXhSSd4u76oyqWd1z3H6uXIyhOJO2_StkAnu9GP-MZQ-W5SLFl_svozYJtiQSel6mE0eMiNJjMq-UfvTElzPipde0XhLQoFBQD3jkKAxE\",\n      \"q\": \"w51sZ56gcxE7xM_bHX8xv8VbcBVHJUtpOZfOmU3mVtBNpV327beqKE-Sy5nTbR0olJg9w3NTYbk_0iSe9SHHAAw_GKiTUiCozogVotV__urdJzTRgaorEqiFixBM3jzq0BkERSNVvt1zT0RQUmEGee6l5pzhHgniEne_2BahTvc\",\n      \"dp\": \"a4msZ9XCRb_SyqujYS3CYkHRfA7zbk3pWnCKqeCG2A7sPHWUn-7KFcLN-NQY9_Ff-oztgg32Db_A4dSoMTV6gvJnwjk_TgXz3cxmNIr4UHjaedplSRtLoM7p_QQ50pittNa90UevtwEnO_PFrxytEJ_KDSAZ0i6-KYB508Ab4RE\",\n      \"dq\": \"Mp6HEd1TidM40zoyalSizAxHGB75dSYpiZTGCpfE9kwa-6LwJh2z85B73SHmO0MjYChxOyWZux0T7c46MBbYhyOkJD5jOQpohGFkr-ltoMqoBeA-Ir6lPMPNIupJ6ex7mvQOTdHR-30UvLthyZ9_7Ieb-Kw5-Or3zG2egE2C2os\",\n      \"qi\": \"LEFTCw84gxciiU8gJXmbG4hD6GTMyP68YRaLoaycmx4vLbBUR3yDuzcgMp9kH52MUrAtXWxessldaXMO-h0pfBboSCJ9VztsswzZ7LDcIo171mSucZbACvqUSyP8BSKF__N2g3wrN5n6ETlUiKbSPtRFnGfC-qiDODGRa4uAMK8\"\n    }\n  ]\n}",
+  "jwksurlcredentials": "{\"clientId\": \"13B06817-BFBCE0B64A1\", \"issuer\": \"http://keycloak.localtest.me:9080/auth/realms/master\", \"tokenEndpoint\": \"http://keycloak.localtest.me:9080/auth/realms/master/protocol/openid-connect/token\"}",
+  "apikey": "Eh1dsMGOYTMMAbh5rsK8L3j0aJHjqKe7",
+  "servicesid": "365e8ab9-79d9-41f6-911a-dbc6b0ee0452",
+  "routesid": "6848fd6a-b966-447d-ad21-4364c53b3b6e"
+}
