+++ e2e/cypress/tests/client-credential-flow/02-client-cred-create-api-prod-auth-pro.ts
@@ -1,10 +1,12 @@
 import HomePage from '../../pageObjects/home'
 import LoginPage from '../../pageObjects/login'
+import NamespaceAccessPage from '../../pageObjects/namespaceAccess'
 import Products from '../../pageObjects/products'
 import ServiceAccountsPage from '../../pageObjects/serviceAccounts'
 import AuthorizationProfile from '../../pageObjects/authProfile'
 
-describe('Create API and Client ID/Secret Auth Profile', () => {
+
+describe('Create API, Product, and Authorization Profiles; Apply Auth Profiles to Product', () => {
   const login = new LoginPage()
   const home = new HomePage()
   const sa = new ServiceAccountsPage()
@@ -22,19 +24,16 @@ describe('Create API and Client ID/Secret Auth Profile', () => {
     cy.fixture('apiowner').as('apiowner')
     cy.visit(login.path)
   })
-
   it('authenticates api owner', () => {
     cy.get('@apiowner').then(({ user }: any) => {
       cy.login(user.credentials.username, user.credentials.password)
     })
   })
-
   it('Activates cc namespace', () => {
     cy.get('@apiowner').then(({ clientCredentials }: any) => {
       home.useNamespace(clientCredentials.namespace)
     })
   })
-
   it('API Owner creates authorization profile for client ID/Secret', () => {
     cy.visit(authProfile.path)
     cy.get('@apiowner').then(({ clientCredentials }: any) => {
@@ -43,7 +42,22 @@ describe('Create API and Client ID/Secret Auth Profile', () => {
       cy.get(authProfile.profileTable).contains(ap.name).should('be.visible')
     })
   })
-
+  it('API Owner creates authorization profile for JWT - Generated Key Pair', () => {
+    cy.visit(authProfile.path)
+    cy.get('@apiowner').then(({ clientCredentials }: any) => {
+      let ap = clientCredentials.jwtKeyPair.authProfile
+      authProfile.createAuthProfile(ap)
+      cy.get(authProfile.profileTable).contains(ap.name).should('be.visible')
+    })
+  })
+  it('API Owner creates authorization profile for JWKS URL', () => {
+    cy.visit(authProfile.path)
+    cy.get('@apiowner').then(({ clientCredentials }: any) => {
+      let ap = clientCredentials.jwks.authProfile
+      authProfile.createAuthProfile(ap)
+      cy.get(authProfile.profileTable).contains(ap.name).should('be.visible')
+    })
+  })
   it('creates a new service account', () => {
     cy.visit(sa.path)
     cy.get('@apiowner').then(({ serviceAccount }: any) => {
@@ -51,7 +65,6 @@ describe('Create API and Client ID/Secret Auth Profile', () => {
     })
     sa.saveServiceAcctCreds()
   })
-
   it('publishes a new API to Kong Gateway', () => {
     cy.get('@apiowner').then(({ clientCredentials }: any) => {
       cy.publishApi('cc-service.yml', clientCredentials.namespace).then(() => {
@@ -61,7 +74,6 @@ describe('Create API and Client ID/Secret Auth Profile', () => {
       })
     })
   })
-
   it('creates a new product in the directory', () => {
     cy.visit(pd.path)
     cy.get('@apiowner').then(({ clientCredentials }: any) => {
@@ -71,7 +83,7 @@ describe('Create API and Client ID/Secret Auth Profile', () => {
       )
     })
   })
-  it('publish product to directory', () => {
+  it('Adds client ID/Secret Env to Client Credentials Test Product', () => {
     cy.visit(pd.path)
     cy.get('@apiowner').then(({ clientCredentials }: any) => {
       let product = clientCredentials.clientIdSecret.product
@@ -80,6 +92,32 @@ describe('Create API and Client ID/Secret Auth Profile', () => {
       product.environment.config.authIssuer = authProfile.name
       product.environment.config.authIssuerEnv = authProfile.environmentConfig.environment
       pd.editProductEnvironmentConfig(product.environment.config)
+    })
+    pd.generateKongPluginConfig('cc-service.yml')
+  })
+  it('Adds Environment for JWT - Generated Key Pair to Client Credentials Test Product', () => {
+    cy.visit(pd.path)
+    cy.get('@apiowner').then(({ clientCredentials }: any) => {
+      let prod = clientCredentials.jwtKeyPair.product
+      let ap = clientCredentials.jwtKeyPair.authProfile
+      pd.addEnvToProduct(prod.name, prod.environment.name)
+      pd.editProductEnvironment(prod.name, prod.environment.name)
+      prod.environment.config.authIssuer = ap.name
+      prod.environment.config.authIssuerEnv = ap.environmentConfig.environment
+      pd.editProductEnvironmentConfig(prod.environment.config)
+    })
+    pd.generateKongPluginConfig('cc-service.yml')
+  })
+  it('API Adds Environment for JWKS URL to Client Credentials Test Product', () => {
+    cy.visit(pd.path)
+    cy.get('@apiowner').then(({ clientCredentials }: any) => {
+      let prod = clientCredentials.jwks.product
+      let ap = clientCredentials.jwks.authProfile
+      pd.addEnvToProduct(prod.name, prod.environment.name)
+      pd.editProductEnvironment(prod.name, prod.environment.name)
+      prod.environment.config.authIssuer = ap.name
+      prod.environment.config.authIssuerEnv = ap.environmentConfig.environment
+      pd.editProductEnvironmentConfig(prod.environment.config)
     })
     pd.generateKongPluginConfig('cc-service.yml')
   })
