+++ e2e/cypress/tests/client-credential-flow/06-jwt-genkp-access-approve-api-rqst.ts
@@ -1,10 +1,53 @@
+import HomePage from '../../pageObjects/home'
+import LoginPage from '../../pageObjects/login'
+import ConsumersPage from '../../pageObjects/consumers'
+
 const njwt = require('njwt')
+
+describe('Access manager approves developer access request for JWT - Generated Key Pair', () => {
+  const home = new HomePage()
+  const login = new LoginPage()
+  const consumers = new ConsumersPage()
+
+  before(() => {
+    cy.visit('/')
+    cy.deleteAllCookies()
+    cy.reload()
+  })
+
+  beforeEach(() => {
+    cy.preserveCookies()
+    cy.fixture('access-manager').as('access-manager')
+    cy.visit(login.path)
+  })
+
+  it('Access Manager logs in', () => {
+    cy.get('@access-manager').then(({ user, clientCredentials }: any) => {
+      cy.login(user.credentials.username, user.credentials.password)
+      home.useNamespace(clientCredentials.namespace)
+    })
+  })
+
+  it('Access Manager approves developer access request', () => {
+    cy.get('@access-manager').then(() => {
+      cy.visit(consumers.path)
+      cy.contains('Review').click()
+      cy.contains('Approve').click()
+    })
+  })
+
+  after(() => {
+    cy.logout()
+    cy.clearLocalStorage({ log: true })
+    cy.deleteAllCookies()
+  })
+})
 
 describe('Make an API request using JWT signed with private key', () => {
   it('Get access token using JWT key pair; make API request', () => {
     cy.readFile('cypress/fixtures/state/store.json').then((store_res) => {
       cy.readFile('cypress/fixtures/state/jwtGenPrivateKey.pem').then((privateKey) => {
-        let jwkCred = JSON.parse(store_res.jwtKeyPairCredentials)
+        let jwkCred = JSON.parse(store_res.jwtkeypaircredentials)
         let clientId = jwkCred.clientId
         let tokenEndpoint = jwkCred.tokenEndpoint
