+++ e2e/cypress/tests/client-credential-flow/08-jwks-url-access-approval-api-rqst.ts
@@ -1,12 +1,54 @@
+import HomePage from '../../pageObjects/home'
 import LoginPage from '../../pageObjects/login'
+import ConsumersPage from '../../pageObjects/consumers'
+
 const njwt = require('njwt')
 const jose = require('node-jose')
+
+describe('Access manager approves developer access request for JWKS URL', () => {
+  const home = new HomePage()
+  const login = new LoginPage()
+  const consumers = new ConsumersPage()
+
+  before(() => {
+    cy.visit('/')
+    cy.deleteAllCookies()
+    cy.reload()
+  })
+
+  beforeEach(() => {
+    cy.preserveCookies()
+    cy.fixture('access-manager').as('access-manager')
+    cy.visit(login.path)
+  })
+
+  it('Access Manager logs in', () => {
+    cy.get('@access-manager').then(({ user, clientCredentials }: any) => {
+      cy.login(user.credentials.username, user.credentials.password)
+      home.useNamespace(clientCredentials.namespace)
+    })
+  })
+
+  it('Access Manager approves developer access request', () => {
+    cy.get('@access-manager').then(() => {
+      cy.visit(consumers.path)
+      cy.contains('Review').click()
+      cy.contains('Approve').click()
+    })
+  })
+
+  after(() => {
+    cy.logout()
+    cy.clearLocalStorage({ log: true })
+    cy.deleteAllCookies()
+  })
+})
 
 describe('Make an API request using JWKS URL flow with JWT signed with generated private key', () => {
   it('Get access token using JWT key pair; make API request', () => {
     cy.readFile('cypress/fixtures/state/store.json').then((store_res) => {
-      let jwkCred = JSON.parse(store_res.jwksUrlCredentials)
-      let ks = store_res.jwksUrlKeys
+      let jwkCred = JSON.parse(store_res.jwksurlcredentials)
+      let ks = store_res.jwksurlkeys
 
       cy.log(ks)
