--- e2e/cypress/tests/client-credential-flow/15-jwks-url-approval.spec.ts
@@ -1,40 +1,0 @@
-import HomePage from '../../pageObjects/home'
-import LoginPage from '../../pageObjects/login'
-import ConsumersPage from '../../pageObjects/consumers'
-
-describe('Access manager approves developer access request for JWKS URL', () => {
-  const home = new HomePage()
-  const login = new LoginPage()
-  const consumers = new ConsumersPage()
-
-  before(() => {
-    cy.visit('/')
-    cy.clearCookies()
-    cy.reload()
-  })
-
-  beforeEach(() => {
-    cy.preserveCookies()
-    cy.fixture('access-manager').as('access-manager')
-    cy.visit(login.path)
-  })
-
-  it('Access Manager logs in', () => {
-    cy.get('@access-manager').then(({ user, clientCredentials }: any) => {
-      cy.login(user.credentials.username, user.credentials.password)
-      home.useNamespace(clientCredentials.namespace)
-    })
-  })
-
-  it('Access Manager approves developer access request', () => {
-    cy.get('@access-manager').then(() => {
-      cy.visit(consumers.path)
-      cy.contains('Review').click()
-      cy.contains('Approve').click()
-    })
-  })
-
-  after(() => {
-    cy.logout()
-  })
-})
