+++ local/cypress-jwks-url/keys.json
@@ -1,1 +1,1 @@
-{"keys":[{"kty":"RSA","kid":"8sxfnMM8IZWpf_dq4Jae7cbYXn0g7nq9HahbV7yQ1Do","use":"sig","alg":"RS256","e":"AQAB","n":"gWjyfx41TvPc536I9M1-36fmc5a1krkVfxFy3zYeQFalv1uHRnIVmoix3i3dzuef3we_hhX5JDFKfkZdXAHSwn1ul_JgsQHJEYBKza0qyG1IOnQ1k6WksStty3iosBxzrh5YfIVQPwrp0ks4QH9zUdilpSZR2I-iQgapVqyuDlv3_w6BpmrxlGhiIcOKyBN6VeXXBJuIS3cT78mIHE7uvS3TeIEErzkMmxOtoHHDhKlUc1yyp0kmSA_C-eO9er52KpKIUTFwVqFUAf3J7vH6fwwGKe7hcVDiOAcv2RIhNOtrG1qTsHoOK1tul_bePEbK4XLhIhrM44lLjOC84ly_tQ"}]}
+{"keys":[{"kty":"RSA","kid":"9EYL-Yl5dqiO0wbCyMKodOI4fC4U4IVaIfkj2kW7cw4","use":"sig","alg":"RS256","e":"AQAB","n":"tYl-DJHVbqClDiNXlsag3FDY92GOG1p9XHdv8Hsjoz9X5lsY4wHCM9d5PEK-Q3Q-cY-YH4w0YbP6D4GqZoRSJbtcsCCgUS5hk2SPNI5Jsv-_RiuzInSrTKLMPgZKDiVoe5W-G5yhoUOmhtABBvrwppnfff5o2JEfdeak1V7zA7WUWJjR6olPxNt3SbAbz456UYMvD7sAW2ap8hDEmYxL_LKOhKb_P4Usoho7BIq0DhT5EoJ7Nnorf3tICfuzMQ6WB53cQaJOV8zKiap6JTQsHKMA5G1gr-onbKfk7om1kYmTJ8UV8or40z70xmeCDNsFPMe3pizh96Q3yXGxeUkjZw"}]}
